ReackQuery frontend structure change for providers and linter
