+++ ReactQuery_FrontEnd/package.json
@@ -24,7 +24,7 @@
     "devDependencies": {
         "@eslint/js": "^9.39.2",
         "@testing-library/jest-dom": "^6.9.1",
-        "@testing-library/react": "^16.3.0",
+        "@testing-library/react": "^16.3.1",
         "@testing-library/user-event": "^14.6.1",
         "@types/react": "^19.2.7",
         "@types/react-dom": "^19.2.3",
@@ -36,7 +36,7 @@
         "eslint-plugin-vitest-globals": "^1.5.0",
         "globals": "^16.5.0",
         "jsdom": "^27.3.0",
-        "vite": "^7.2.7",
-        "vitest": "^4.0.15"
+        "vite": "^7.3.0",
+        "vitest": "^4.0.16"
     }
 }
