+++ ReactQuery_FrontEnd/src/App.jsx
@@ -1,22 +1,22 @@
 import { Routes, Route, useMatch, Navigate, } from 'react-router-dom'
-import { setNotification, clearNotification, } from './context/notifications'
-import { setUser, } from './context/user'
+import { setNotification, clearNotification, } from './contexts/notification'
+import { setUser,  } from './contexts/user'
 import { useContext, } from 'react'
 import { useEffect, } from 'react'
 import { useQuery, useQueryClient, useMutation, } from '@tanstack/react-query'
 import AccountForm from './components/AccountForm'
 import Blog from './components/Blog'
-import Blogs from './components/Blogs'
 import BlogService from './services/blogsService'
+import Blogs from './components/Blogs'
 import CreateForm from './components/CreateForm'
 import Message from './components/Message'
-import notificationContext from './context/notifications'
+import Stack from 'react-bootstrap/Stack'
 import Toggle from './components/Toggle'
 import User from './components/User'
-import userContext from './context/user'
 import Users from './components/Users'
+import notificationContext from './contexts/notification'
+import userContext from './contexts/user'
 import userService from './services/userService'
-import Stack from 'react-bootstrap/Stack'
 
 const App = () => {
     const [notification, notificationDispatch,] = useContext(notificationContext,)
