+++ ReactQuery_FrontEnd/src/main.jsx
@@ -1,8 +1,8 @@
 import ReactDOM from 'react-dom/client'
 import App from './App'
 import { QueryClientProvider, QueryClient, } from '@tanstack/react-query'
-import { NotificationContextProvider, } from './context/notifications'
-import { UserContextProvider, } from './context/user'
+import NotificationContextProvider from './providers/notifications'
+import UserContextProvider from './providers/user'
 import { BrowserRouter as Router, } from 'react-router-dom'
 
 const queryClient = new QueryClient()
