+++ Redux_FrontEnd/package.json
@@ -11,7 +11,7 @@
         "preview": "vite preview"
     },
     "dependencies": {
-        "@reduxjs/toolkit": "^2.11.1",
+        "@reduxjs/toolkit": "^2.11.2",
         "@stylistic/eslint-plugin": "^5.6.1",
         "axios": "^1.13.2",
         "bootstrap": "^5.3.8",
@@ -26,7 +26,7 @@
     "devDependencies": {
         "@eslint/js": "^9.39.2",
         "@testing-library/jest-dom": "^6.9.1",
-        "@testing-library/react": "^16.3.0",
+        "@testing-library/react": "^16.3.1",
         "@testing-library/user-event": "^14.6.1",
         "@types/react": "^19.2.7",
         "@types/react-dom": "^19.2.3",
@@ -38,7 +38,7 @@
         "eslint-plugin-vitest-globals": "^1.5.0",
         "globals": "^16.5.0",
         "jsdom": "^27.3.0",
-        "vite": "^7.2.7",
-        "vitest": "^4.0.15"
+        "vite": "^7.3.0",
+        "vitest": "^4.0.16"
     }
 }
